Isolate NoiseModel tests
